forget DRY, we're going DAMP
