--- task-sdk/tests/task_sdk/definitions/utils/test_deadlines.py
@@ -1,39 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-from __future__ import annotations
-
-import pytest
-from task_sdk.definitions.test_dag import DEFAULT_DATE
-
-from airflow.sdk.definitions.deadline import DeadlineReference
-
-DAG_ID = "dag_id_1"
-
-TEST_CALLBACK_PATH = f"{__name__}.test_callback_for_deadline"
-TEST_CALLBACK_KWARGS = {"arg1": "value1"}
-
-
-def test_callback_for_deadline():
-    """Used in a number of tests to confirm that Deadlines and DeadlineAlerts function correctly."""
-    pass
-
-
-REFERENCE_TYPES = [
-    pytest.param(DeadlineReference.DAGRUN_LOGICAL_DATE, id="logical_date"),
-    pytest.param(DeadlineReference.DAGRUN_QUEUED_AT, id="queued_at"),
-    pytest.param(DeadlineReference.FIXED_DATETIME(DEFAULT_DATE), id="fixed_deadline"),
-]
